adds language specific response data
